move recordAnswer action to universal file, and update imports
