+++ src/applications/check-in/actions/universal/index.js
@@ -6,3 +6,12 @@ export const setApp = application => {
     payload: { app: application },
   };
 };
+
+export const RECORD_ANSWER = 'RECORD_ANSWER';
+
+export const recordAnswer = answer => {
+  return {
+    type: RECORD_ANSWER,
+    payload: answer,
+  };
+};
