+++ src/applications/check-in/actions/universal/universal.actions.unit.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import { setApp, SET_APP } from './index';
+import { setApp, SET_APP, recordAnswer, RECORD_ANSWER } from './index';
 
 describe('check-in', () => {
   describe('universal actions', () => {
@@ -12,6 +12,18 @@ describe('check-in', () => {
       it('should return correct structure', () => {
         const action = setApp('preCheckIn');
         expect(action.payload.app).equal('preCheckIn');
+      });
+    });
+    describe('recordAnswer', () => {
+      it('should return correct action', () => {
+        const action = recordAnswer({});
+        expect(action.type).to.equal(RECORD_ANSWER);
+      });
+      it('should return correct structure', () => {
+        const action = recordAnswer({
+          demographicsUpToDate: 'yes',
+        });
+        expect(action.payload.demographicsUpToDate).equal('yes');
       });
     });
   });
