+++ src/applications/check-in/day-of/pages/Demographics.jsx
@@ -6,7 +6,8 @@ import recordEvent from 'platform/monitoring/record-event';
 import { useFormRouting } from '../../hooks/useFormRouting';
 import BackToHome from '../../components/BackToHome';
 import Footer from '../../components/layout/Footer';
-import { recordAnswer, seeStaffMessageUpdated } from '../../actions/day-of';
+import { seeStaffMessageUpdated } from '../../actions/day-of';
+import { recordAnswer } from '../../actions/universal';
 import DemographicsDisplay from '../../components/pages/demographics/DemographicsDisplay';
 import { makeSelectVeteranData } from '../../selectors';
