+++ src/applications/check-in/day-of/pages/EmergencyContact.jsx
@@ -8,7 +8,8 @@ import { useFormRouting } from '../../hooks/useFormRouting';
 import BackButton from '../../components/BackButton';
 import BackToHome from '../../components/BackToHome';
 import Footer from '../../components/layout/Footer';
-import { recordAnswer, seeStaffMessageUpdated } from '../../actions/day-of';
+import { seeStaffMessageUpdated } from '../../actions/day-of';
+import { recordAnswer } from '../../actions/universal';
 import EmergencyContactDisplay from '../../components/pages/emergencyContact/EmergencyContactDisplay';
 import { makeSelectVeteranData } from '../../selectors';
