+++ src/applications/check-in/day-of/pages/NextOfKin.jsx
@@ -8,7 +8,8 @@ import { useFormRouting } from '../../hooks/useFormRouting';
 import BackButton from '../../components/BackButton';
 import BackToHome from '../../components/BackToHome';
 import Footer from '../../components/layout/Footer';
-import { recordAnswer, seeStaffMessageUpdated } from '../../actions/day-of';
+import { seeStaffMessageUpdated } from '../../actions/day-of';
+import { recordAnswer } from '../../actions/universal';
 import NextOfKinDisplay from '../../components/pages/nextOfKin/NextOfKinDisplay';
 import { makeSelectVeteranData } from '../../selectors';
 import { URLS } from '../../utils/navigation';
